ovii_backend/requirements.txt
asgiref==3.8.1
celery==5.4.0
channels==4.0.0
channels-redis==4.2.0
Django==5.0.7
django-cors-headers==4.4.0
django-phonenumber-field==7.3.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
phonenumbers==8.13.40
PyJWT==2.8.0
redis==5.0.7
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1